requirements_full_backup.txt
# Django Backend Requirements - Rejlers API System
# Core Django Framework
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==42.0.2
bcrypt==4.1.2

# API Documentation
drf-spectacular==0.27.1
django-silk==5.0.4
django-filter==23.5

# File Handling & Storage
django-storages==1.14.2
boto3==1.34.44
botocore==1.34.44

# AWS Services
django-ses==3.5.2
watchtower==3.0.1

# Caching & Performance
redis==5.0.1
django-redis==5.4.0
celery==5.3.6

# Monitoring & Logging
django-extensions==3.2.3
sentry-sdk==1.40.6

# Testing
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0

# Development Tools
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0

# AI & Machine Learning
openai==1.12.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.0
Pillow==10.2.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1